retro/requirements2.txt
## gooey-gpu
Pillow ~= 9.4.0
sentry-sdk ~= 1.14.0
celery ~= 5.3.0
pydantic ~= 1.10.9
redis ~= 4.5.5

## xformers
xformers==0.0.16
triton==2.0.0.dev20221202

## huggingface diffusers
diffusers ~= 0.14.0
transformers ~= 4.26.1
accelerate ~= 0.17.0
sentencepiece ~= 0.1.97

## wav2lip
librosa ~= 0.7.0
opencv-python ~= 4.7.0.68
tqdm ~= 4.45.0
numba == 0.48
mediapipe ~= 0.10.1
scipy ~= 1.10.1

## gfpgan
git+https://github.com/xinntao/Real-ESRGAN.git
git+https://github.com/TencentARC/GFPGAN.git
git+https://github.com/xinntao/BasicSR
git+https://github.com/xinntao/facexlib
